refactor: amalgamated all orderbook related hooks
